move inner class to independent file
